shelly-irrigation-control/requirements.txt
requests
python-dotenv
rich
pytz